+++ index.ts
@@ -1,5 +1,5 @@
-import { httpServer } from './src/http_server/index.js';
-import { createWsServer } from './src/ws_server/index.js';
+import { httpServer } from './src/http_server/index';
+import { createWsServer } from './src/ws_server/index';
 
 const HTTP_PORT = 8181;
