+++ package.json
@@ -15,6 +15,7 @@
     "license": "ISC",
     "devDependencies": {
         "@types/node": "^18.11.18",
+        "@types/ws": "^8.5.4",
         "nodemon": "^2.0.20",
         "ts-node": "^10.9.1",
         "typescript": "^4.9.4"
