+++ src/ws_server/commands-handler.ts
@@ -1,8 +1,12 @@
+import { WebSocket } from "ws";
+import Ws from "ws";
+import { Command } from "./commands/interfaces/command";
+
 const UNKNOWN_COMMAND_MESSAGE = 'Unknown command';
 const OPERATION_FAILED_MESSAGE = 'Operation failed';
-const commandsMap = {};
+const commandsMap: { [name: string]: Command } = {};
 
-const handleCommandMessage = async (commandMessage, sendAnswer) => {
+const handleCommandMessage = async (commandMessage: Ws.RawData, sendAnswer: (data: any, cb?: (err?: Error) => void) => void) => {
     const commandParts = commandMessage.toString().split(' ');
     const commandName = commandParts[0].trim().toLocaleLowerCase();
     const command = commandsMap[commandName];
@@ -20,7 +24,7 @@ const handleCommandMessage = async (commandMessage, sendAnswer) => {
     }
 };
 
-const registerCommands = (commands) => commands.forEach(c => commandsMap[c.commandName] = c );
+const registerCommands = (commands: Command[]) => commands.forEach(c => commandsMap[c.commandName] = c );
 
 export default {
     registerCommands,
