+++ src/ws_server/commands/draw_square.ts
@@ -1,6 +1,7 @@
-import { drawRectangleCommand } from './draw_rectangle.js';
+import { drawRectangleCommand } from './draw_rectangle';
+import { Command } from './interfaces/command';
 
-const drawSquareCommand = {
+const drawSquareCommand: Command = {
     commandName: 'draw_square',
     run: async (args, sendResult) => {
         const size = Number(args[0]);
