+++ src/ws_server/commands/interfaces/command.ts
@@ -1,0 +1,7 @@
+export interface Command {
+    commandName: string;
+    run: (
+        args: (string | Number)[],
+        sendResult: (data: any, cb?: (err?: Error) => void) => void
+    ) => void;
+}
