+++ src/ws_server/commands/mouse-down.ts
@@ -1,6 +1,7 @@
 import { mouse, down } from '@nut-tree/nut-js';
+import { Command } from './interfaces/command';
 
-const mouseDownCommand = {
+const mouseDownCommand: Command = {
     commandName: 'mouse_down',
     run: async (args, sendResult) => {
         const offset = Number(args[0]);
