+++ src/ws_server/commands/mouse-left.ts
@@ -1,6 +1,7 @@
 import { mouse, left } from '@nut-tree/nut-js';
+import { Command } from './interfaces/command';
 
-const mouseLeftCommand = {
+const mouseLeftCommand: Command = {
     commandName: 'mouse_left',
     run: async (args, sendResult) => {
         const offset = Number(args[0]);
