+++ src/ws_server/commands/mouse-position.ts
@@ -1,6 +1,7 @@
 import { mouse, up } from '@nut-tree/nut-js';
+import { Command } from './interfaces/command';
 
-const mousePositionCommand = {
+const mousePositionCommand: Command = {
     commandName: 'mouse_position',
     run: async (args, sendResult) => {
         const position = await mouse.getPosition();
