+++ src/ws_server/commands/mouse-right.ts
@@ -1,6 +1,7 @@
 import { mouse, right } from '@nut-tree/nut-js';
+import { Command } from './interfaces/command';
 
-const mouseRightCommand = {
+const mouseRightCommand: Command = {
     commandName: 'mouse_right',
     run: async (args, sendResult) => {
         const offset = Number(args[0]);
