+++ src/ws_server/commands/mouse-up.ts
@@ -1,6 +1,7 @@
 import { mouse, up } from '@nut-tree/nut-js';
+import { Command } from './interfaces/command';
 
-const mouseUpCommand = {
+const mouseUpCommand: Command = {
     commandName: 'mouse_up',
     run: async (args, sendResult) => {
         const offset = Number(args[0]);
