+++ tsconfig.json
@@ -7,8 +7,8 @@
       "lib": ["es6"],
       "resolveJsonModule": true,
       "esModuleInterop": true,
-      // "strict": true,
-      // "noImplicitAny": true,
+      "strict": true,
+      "noImplicitAny": true,
       "skipLibCheck": true
     },
     "include": ["./src/**/*", "index.ts"]
